% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffit.R
\name{ffit}
\alias{ffit}
\title{Discrete Fourier Series fit}
\usage{
ffit(y, fit.order = 10, low.xlim = 1, up.xlim = length(y))
}
\arguments{
\item{\code{y}}{Numeric data.}

\item{\code{low.xlim}}{First x position. Default: \code{1}.}

\item{\code{up.xlim}}{Last x position. Default: \code{length(y)}}
}
\value{
\code{fit.order} Model order. Default: \code{10}
}
\description{
Discrete Fourier Series fit
}
\examples{
# Senoid
set.seed(123)
t0 <- 1:(5*24*60)
y <- ts(15 + 0.001*t0 + 10*sin(2*pi*t0/(length(t0)/5)) + rnorm(length(t0)), freq=length(t0)/5)
plot(y, type='l')
fit <- ffit(y)
plot(ft, lwd = 3, col = 'red', xlim = range(1:length(y)), main = paste0('Fit order ', fit$fit.order))
points(1:length(y), y, pch = '.')

# Random data
n <- 1000
set.seed(4321)
y2 <- cumsum(rnorm(n))
(fit <- ffit(y2, fit.order = 100))
plot(ft, lwd = 3, col = 'red', xlim = range(1:n), main = paste0('Fit order ', fit$fit.order))
points(1:n, y2, pch = '.')
}
