% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft.R
\name{ft}
\alias{ft}
\title{Plot to Discrete Fourier Series fit}
\usage{
ft(x)
}
\arguments{
\item{\code{x}}{Output from \code{dfs} function in an object called \code{fit}.}
}
\description{
Plot to Discrete Fourier Series fit
}
\examples{
n <- 1000000
(fit <- ffit(rnorm(n)))
plot(ft, lwd = 3, col = 'red', xlim = range(1:n), main = paste0('Fit order ', 10))
}
